requirements.txt
fastapi==0.124.0
pytest==9.0.2
pytest-xdist==3.8.0
requests==2.32.5
allure-pytest==2.15.2
playwright==1.56.0
curlify==3.0.0
python-dotenv==1.2.1
voluptuous==0.15.2
pytest-voluptuous==1.2.0
uvicorn==0.38.0